3.4.2 Reorganization of classes, added Jane Figure
