+++ src/main/java/edu/kis/powp/jobs2d/drivers/adapter/AbstractDriverAdapter.java
@@ -23,7 +23,7 @@ public class AbstractDriverAdapter extends AbstractDriver {
         line.setEndCoordinates(x, y);
 
         drawPanelController.drawLine(line);
-        
+
         setPosition(x, y);
     }
