+++ src/test/java/edu/kis/powp/jobs2d/events/SelectTestFigureJoeOptionListener.java
@@ -6,11 +6,11 @@ import java.awt.event.ActionListener;
 import edu.kis.powp.jobs2d.drivers.DriverManager;
 import edu.kis.powp.jobs2d.magicpresets.FiguresJoe;
 
-public class SelectTestFigureOptionListener implements ActionListener {
+public class SelectTestFigureJoeOptionListener implements ActionListener {
 
 	private DriverManager driverManager;
 
-	public SelectTestFigureOptionListener(DriverManager driverManager) {
+	public SelectTestFigureJoeOptionListener(DriverManager driverManager) {
 		this.driverManager = driverManager;
 	}
